--- app/(base)/profile/[nickname]/components/CompletedArenaList.tsx
@@ -1,68 +1,0 @@
-"use client";
-
-import { useEffect, useState } from "react";
-import useArenas from "@/hooks/useArenas";
-import CompleteArenaCard from "@/app/(base)/arenas/components/CompleteArenaCard";
-import Pager from "@/app/components/Pager";
-
-export default function CompletedArenaList({ memberId }: { memberId: string }) {
-    const [currentPage, setCurrentPage] = useState(1);
-    const pageSize = 6;
-    const status = 5; // 종료된 투기장
-
-    const { arenaListDto, loading, error } = useArenas({
-        currentPage,
-        status,
-        mine: true,
-        pageSize,
-        targetMemberId: memberId,
-    });
-
-    useEffect(() => {
-        if (!loading && arenaListDto?.arenas) {
-            console.log("✅ 완료된 투기장 개수:", arenaListDto.arenas.length);
-        }
-    }, [loading, arenaListDto]);
-
-    if (loading) {
-        return <p className="text-sm text-font-200">로딩 중입니다...</p>;
-    }
-
-    if (error) {
-        return (
-            <p className="text-sm text-red-500">
-                투기장 정보를 불러오는 데 실패했습니다.
-            </p>
-        );
-    }
-
-    if (!arenaListDto || arenaListDto.arenas.length === 0) {
-        return (
-            <p className="text-sm text-font-200">
-                참여한 종료된 투기장이 없습니다.
-            </p>
-        );
-    }
-
-    return (
-        <div className="flex w-full flex-col items-center gap-6">
-            <div className="grid grid-cols-1 gap-6 px-1 md:grid-cols-2">
-                {arenaListDto.arenas.map((arena) => (
-                    <CompleteArenaCard
-                        key={arena.id}
-                        {...arena}
-                        showBadgeIconOnly={true}
-                    />
-                ))}
-            </div>
-
-            {/* 페이지네이션 */}
-            <Pager
-                currentPage={arenaListDto.currentPage}
-                endPage={arenaListDto.endPage}
-                pages={arenaListDto.pages}
-                onPageChange={setCurrentPage}
-            />
-        </div>
-    );
-}
