+++ app/(base)/profile/components/RecruitingArenaList.tsx
@@ -2,24 +2,28 @@
 
 import { useEffect, useState } from "react";
 import useFetchArenas from "@/hooks/useArenas";
-import VotingArenaCard from "@/app/(base)/arenas/components/VotingArenaCard";
+import RecruitingArenaCard from "@/app/(base)/arenas/components/RecruitingArenaCard";
 import Pager from "@/app/components/Pager";
 
-export default function VotingArenaList({ memberId }: { memberId: string }) {
+type ArenaListProps = {
+    memberId?: string;
+};
+
+export default function RecruitingArenaList({ memberId }: ArenaListProps) {
     const [currentPage, setCurrentPage] = useState(1);
     const pageSize = 6;
 
     const { arenaListDto, loading, error } = useFetchArenas({
         currentPage,
-        status: 4, // 상태 4 = 투표 중
-        mine: true, // 내가 생성하거나 참여한 투기장만
+        status: 1,
+        mine: true,
         pageSize,
-        targetMemberId: memberId,
+        ...(memberId ? { targetMemberId: memberId } : {}),
     });
 
     useEffect(() => {
         if (!loading && arenaListDto?.arenas) {
-            console.log("✅ 투표 중 투기장 개수:", arenaListDto.arenas.length);
+            console.log("✅ 모집 중 투기장 개수:", arenaListDto.arenas.length);
         }
     }, [loading, arenaListDto]);
 
@@ -38,7 +42,7 @@ export default function VotingArenaList({ memberId }: { memberId: string }) {
     if (!arenaListDto || arenaListDto.arenas.length === 0) {
         return (
             <p className="text-sm text-font-200">
-                투표 중인 투기장이 없습니다.
+                모집 중인 투기장이 없습니다.
             </p>
         );
     }
@@ -48,23 +52,15 @@ export default function VotingArenaList({ memberId }: { memberId: string }) {
             <div className="w-full">
                 <div className="grid grid-cols-1 gap-6 px-1 md:grid-cols-2">
                     {arenaListDto.arenas.map((arena) => (
-                        <VotingArenaCard
+                        <RecruitingArenaCard
                             key={arena.id}
-                            id={arena.id}
-                            title={arena.title}
-                            creatorNickname={arena.creatorNickname}
-                            creatorScore={arena.creatorScore}
-                            challengerNickname={arena.challengerNickname}
-                            challengerScore={arena.challengerScore}
-                            voteEndDate={new Date(arena.voteEndDate)}
-                            voteCount={arena.voteCount}
+                            {...arena}
+                            startDate={new Date(arena.startDate)}
                             showBadgeIconOnly={true}
                         />
                     ))}
                 </div>
             </div>
-
-            {/* 페이지네이션 */}
             <Pager
                 currentPage={arenaListDto.currentPage}
                 endPage={arenaListDto.endPage}
